Implement -functions and filter of test.
The normal qtest framework allows you to list the available tests using
the "-functions" switch, and then to pass the name of a particular
function or functions so as to restrict the run to only those test
functions.  This is very very important for fixing test breakages
because otherwise the debug output from other tests pollutes the console
so badly that it becomes difficult to progress.

Change-Id: Iac0bc663d406f65874ce4bc1be0f816b6b7fa80c