QTBUG-19248 - fix uninstaller for win
This was a real pain to fix.  On Win 7 files written to the users home start
menu directory were being relocated to a system directory.  For now the
uninstaller also checks in that location and deletes the start links from
there as well.

Also since binaries cannot all be put into Qt's bin directory, alongside the
libraries, due to potential name clashes, a batch file had to be used to set
the path and then launch the start program.  This means that uninstalling
can also be done more easily by just deleting the quick3d\bin directory along
with its contents.