Make this a bit nicer as the welcome app.
Windows installers need an app to launch after the installer runs, and to
carry the registration key.  Can do without it, but its a pain.

This app is now the qglinfo app so make it a bit more presentable.

Also make it show a welcome message the first time its run.