Fix QTBUG-19249
Other fixes to windows QML also needed, but this one was package related.
Have setup to launch the sample programs by using windows shortcut (.lnk)
files as these provide the ability to specify the current directory to
run the program from, making it possible to locate the .dll's.

Using shortcut files implies also have to create icons for all the
programs as well.