New troubleshooing, and Fix qdoc warnings.
Add some trouble-shooting tips around the Qt Declarative case detection
on windows.

It seems now that qdoc3 has coded inside it some reference to an "All
Modules" page, and a few other pages, which relate to the breadcrumbs
navigation.

These pages get created as part of the Qt Doc, but while the Qt3D doc is
standalone, these pages do not exist and as a result a pile of warnings
spew out when running qdoc.

It might be that when Qt3D is made a full module, and the qt5 guys
finally fix doc generation (AIUI its borken now) that this change has to
be backed out, but for now it means that qdoc3 runs without generating
so much error output that real errors cannot be seen.

Change-Id: I1d689339e5a128991ed10922c729c45f58561eb4