Refactor packaging.
Packaging was always not so good. With the arrival of yet another
packaging system to support, with QtSDK and Harmattan, it was overdue
time to refactor the build system.

Factored out all the places where Qt3D and QtQuick3D dependencies were
exactly 3 levels below the source .pro file. This catches all of the
examples and all of the demos, except for the ones in the tutorials
directories. It is fine if those don't appear in packages, since their
main value is in examining them in the IDE, rather than just having a
package run them for you. It also catches both of the import plugins.

Thus we save a huge amount of duplication of packaging information and
can place the Qt SDK Harmattan packaging into a couple of variables in a
new pkg.pri file. The two deps are now in a qt3d_pkg_dep.pri and
qt3dquick_pkg_dep.pri file which include the pkg.pri.

Change-Id: I28de17d7faa5ccf9157afad99471079c0b62ee87
Reviewed-on: http://codereview.qt.nokia.com/1191
Reviewed-by: Sarah Jane Smith