First round of fixing QTBUG-18823
Fixing reported problems with the isCullable function.

Implement an improved very clear unit test for the
isCullable functions, that are split out by value
rather than being buried in a for loop.

Also spend some time documenting the expected beha-
vior of these functions and visualising this.

The isCullableVert test has been marked with

QSKIP(..., SkipAll)

which means that CI should let it pass so we can
integrate and keep moving on with a fix for the bug.

Change-Id: Ibb8c9493256cb1480c2bbda1792f3751830031cd
Reviewed-on: http://codereview.qt.nokia.com/1928
Reviewed-by: Julian de Bhal <julian.debhal@nokia.com>