Fix QTBUG-18823.
Implemented correctly the clip space algorithm and
fixed a number of minor but fatal problems with the
previous implementation.

Also added useful debug to the camera class, which
helped during debugging.

Made a correction to the frustum doc, which impacted
the testing work, and added a section on the aspect
ratio which had garnered some questions on the list.

In the tests, added some tests that cover the view
frustum being forwards or backwards over the object
being viewed/culled, which was not covered before.

Change-Id: I488441eb89b5b112daa973d13260880ee3286038
Reviewed-on: http://codereview.qt.nokia.com/2258
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Danny Pope <daniel.pope@nokia.com>
Reviewed-by: Zheng Liu <zheng.g.liu@nokia.com>
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>