Add signals for culling.
These are useful for testing but also for applications that want to
respond to an item becoming visible or going out of view.

Tests using this to be added in seperate commit.

Reviewed-on: http://codereview.qt.nokia.com/2760
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Patrick Burke <patrick.burke@nokia.com>

Create the node property.

Earlier commits added the culling signal to scene nodes.  Here are a
property for accessing the scene node so that the culling signal
actually can be used from within a qml program.

Change-Id: Ica74b51166eb8a63ad00ffcc4aab090816733985
Reviewed-on: http://codereview.qt.nokia.com/2922
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Danny Pope <daniel.pope@nokia.com>
Reviewed-by: Patrick Burke <patrick.burke@nokia.com>
Reviewed-on: http://codereview.qt.nokia.com/3252
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>