Put the skybox stuff into QML3D.
This skybox functionality was originally written for the photobrowser3d
demo.  Begin creating a flick3d demo which will eventually replace the
photobrowser3d using the network aware url support for textures, and the
3d data models recently implemented.

At present all the flickr3d demo does is to display a textured cube
inside a skybox.

Change-Id: I1d8bbb4a8853b1e39de96d9eab220b198df31707
Reviewed-on: http://codereview.qt.nokia.com/2921
Reviewed-by: Patrick Burke <patrick.burke@nokia.com>
Reviewed-on: http://codereview.qt.nokia.com/3249
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>