Put the skybox stuff into QML3D.
This skybox functionality was originally written for the photobrowser3d
demo.  Begin creating a flick3d demo which will eventually replace the
photobrowser3d using the network aware url support for textures, and the
3d data models recently implemented.

At present all the flickr3d demo does is to display a textured cube
inside a skybox.

Change-Id: Ibb92da172a9ee3996da20859e7831d4247bf7b01
Reviewed-on: http://codereview.qt.nokia.com/2921
Reviewed-by: Patrick Burke <patrick.burke@nokia.com>