Doc fixes for QTBUG-20784
This bug report relates to problems experienced with QGLDome and
QGLCylinder.  These classes are quite well documented, except for the
node structure that they produce in the scene, as they interact with
QGLBuilder.

It makes no sense to document this structure as it is subject to change
as an implementation detail.  Instead the right answer is to use the
qDumpScene function (or other development techniques) to examine the
structure that results, and to then use that structure to code for the
desired scene rendering.

Have added documentation to get this message across.

Change-Id: If9ace96dcf8cbf54658b2d630481f9eebd39ae93
Reviewed-on: http://codereview.qt.nokia.com/3248
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>