Make work with threads.
This current design was not good, as it made the drawItem change the
structure of the geometry potentially every frame.

If someone does actually want to animate the lod (gakk) then this is
going to pull a lot of triangles.

But anyway - don't do it in the rendering code.  This should always be
done in response to the slot that updates the LOD.  In the threaded case
this is essential as it would result in the new QGLSceneNode objects
being located in the render thread and thus they cannot be parented to
the sphere - leak at best, crash at worst.

Change-Id: Icf5dd4605d622897d231e50d7b828a9f1ad04815
Reviewed-on: http://codereview.qt.nokia.com/3836
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Patrick Burke <patrick.burke@nokia.com>