Picking implementation for Qt5/QML2.
Main issues at present are that lander does not work, and the picking
areas are out.  Bugs will be entered for this.

With picking in cases where zooming is done the picking is out a larger
amount.  Most examples do not seem to show this, but zooming the
tst_basic example does show this.  Also the tea_service example works,
except the teaspoons cannot be clicked - probably due to this issue.

The lander issue is with somehow the MouseArea not getting the click.
The QSGItem::mouseEvent() call does not seem to have the effect of
passing the click through to the MouseArea item.  The logic in
viewport.cpp is correct, but the event delivery is not happening.

Change-Id: I7ae8a037437bc7931ed71b79d805471836af9f94
Reviewed-on: http://codereview.qt-project.org/4706
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Patrick Burke <patrick.burke@nokia.com>