Fixes for surface vs QOpenGLContext & QWindow.
With the qt5 refactor branch going in the ability to get a QPaintDevice
has gone away - this was really a "relic from the desktop days".

Painting still applies for FBO's and so on, but in the case of QML we
are on an SG view and there is no paint device there.

So move the QGLView to using a QWindow and QOpenGLContext as well to
follow suit and replace all the logic in surface stuff to cope with
that.

Also update all examples, demos, tutorials and tests.

Change-Id: Ie8dbeb97c87ef0821326fb7ccf5d5d4b1f90fd06
Reviewed-on: http://codereview.qt-project.org/5900
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>