Bring depth buffer fix to viewport as well.
Recent post-refactor fixes in QGLViewport were required because the
context and window were not getting a depth buffer.

Now this problem is manifesting in Viewport as well.  Fix by grabbing
the canvas as soon as it is instantiated and setting a depth buffer onto
its format.  This should get carried over into the QOpenGLContext when
it is created on the canvas.

Change-Id: Ib21966b1f84ffa7baa7fc33b996f940380ca8293
Reviewed-on: http://codereview.qt-project.org/6408
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Patrick Burke <patrick.burke@nokia.com>