Swap always.
For now elide the logic that asks what swap behavior we are supposed to
be able to do, and just try to swap anyway.

Looks as though the report of the format we get is not quite accurate.

Dig further in subsequent commits but get this in for now so we can at
least display our C++ examples.

Change-Id: Id2b0dcf69490e90482d925370263485c7b555d9d
Reviewed-on: http://codereview.qt-project.org/6468
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Patrick Burke <patrick.burke@nokia.com>