Fix skating viewport, and quit.
Under qt 5 what seems to have been a long standing bug was exposed
where glViewport was called with the first two arguments set to
the x, y origin of the window.  These have always been 0, 0 apparently
but under xcb that is not the case and dragging the window of a running
qt3d app across the desktop was causing the gl to scroll across inside
the window.  Looked pretty weird.  Now set to 0, 0 as elsewhere and seems
to have fixed the problem.

Also fixed quit not working, and put a test for the new quit enabled
viewport into basket app.  Will need to be added to other apps also.

Change-Id: I85651cc0df2fdbef466f843e12c046986f495b29
Reviewed-on: http://codereview.qt-project.org/6505
Reviewed-by: Patrick Burke <patrick.burke@nokia.com>