Get qml infrastructure in place again.
After recent fixes to remove crufty build logic, there was a problem
where the qml plugins were installed, but not the qml infrastructure
files that those plugins required to work: qmldir and friends.

These would have gotten installed via a "make install" but it makes no
sense to have a situation where one part is installed but it is
non-functional because the rest of the required files are not present.

Change-Id: Ifd01539d3ba12380282ec9ad8ba37388ba772a06
Reviewed-by: Patrick Burke <patrick.burke@nokia.com>