Make standalone Item3Ds work.
This is kind of a hack but its so succinct and seems to work so well
that I'm inclined to give it a run and see if there is any problems
with it before going for something much more complex and disruptive.

The alternative would be to try to factor out the rendering logic
from the Viewport class into a new class in the libQt3DQuick.so - where
it can be accessed by the Item3D implementation (QDeclarativeItem3D)
and to make a basic Viewport QML item that used the rendering logic, and to
have Item3D use it if it was not a descendant of a Viewport.

The hack is good because it avoids all this hand-wringing and
anxiety by simply asking the declarative system to create us a viewport using a
piece of script, so we don't even need to reference the viewport
class at all.

Change-Id: I16b4d4666c86e679dd4b9d81c54a8cd12f4a957a
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>