Make a solar system example in qml.
This is just the first pass.  The shaders do not seem to be working for
some reason, so some checking there.

Another motivation for this example is to explore QTBUG-22395
which relates to the qml sphere being duplicated every time you invoke
it even if you make a component out of it.  So in a scene with a
10x10x10 "salt crystal" made out of spheres, you have 1000 copies of the
same geometry.

Change-Id: Id0c3afb269b75ec00530d2f96495540085cf6c42
Reviewed-by: Patrick Burke <patrick.burke@nokia.com>