Remove the tga plugin.
At some point Qt has gained support for tga.  Even tho the documentation
on QImageReader::supportedImageFormats() says that tga is not supported
and a google search says the same thing; writing a C++ program that
outputs the result of the supportedImageFormats() function shows that
tga is supported, and creating a QML test program that loads a tga
shows that it does as well.

It looks as tho the support comes from libjpeg, and at some point has
been enabled, without any change in doco.  When this plugin was first
written there definitely was no support.

Time to retire this clunky hack in favor of the supported one in Qt.

Change-Id: I94e2107ff8c662b80fa27818b070b6b54c443101
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>