Fixing harmattan pkgs
The fixes are going into the qtc_harmattan packaging scripts, since the
scratchbox for harmattan seems to be no longer working.  Eventually we
should remove it and put the qtc_harmattan one in instead, but for now
keep both around until its proven to work.  This is also important
because bifh might need it.

Basically - remove the translation line, fix up a lot of broken desktop
files, use a different location for the icons, and generally fix all
the problems as described in the bug.

Task-number: QTBUG-21955

Change-Id: I02cdf6c02ef9515b14384aaaf8e039a09a6cb82b
Reviewed-by: Patrick Burke <patrick.burke@nokia.com>