Improve qDumpScene.
The output of this useful debugging tool was a bit erratic, with
spurious newlines, and hard to read formatting.  Fix, that and also
while on the job improve formatting, and fix the fact that it became
undefined with QT_NO_DEBUG_STREAM which could result in programs
breaking when defines were changed.  Now if QT_NO_DEBUG_STREAM is
defined the function is still there, but just exits without doing
anything.  Also add a non-detailed mode for quick summaries, and for
complex scenes where too much detail would be produced.

Add a debugging flag to the viewport.  Get a dump of the scene when you
set a showSceneGraph flag in your qml.  This could be improved as it
does not handle cases where the Item3D's are in components.  Its a good
start.

Change-Id: I8a5ef3e8739dbe72617aad33377c691838a06fa2
Reviewed-by: Patrick Burke <patrick.burke@nokia.com>