Performance optimizations for graphics scale.
When animating or otherwise setting values, it is common to set the
origin for the scale to zero, and the value of the scale to 1.0.  Both
of these mean that there is no net effect when the transform is applied,
but several matrix multiplications would be done anyway.

Optimisations for these common cases have been added so it is now
possible to set a transform for scale to 1, 1, 1 and not have it cost.

Change-Id: I63b2ff39303cd20e23cb44a5615471f3da7f59b9
Reviewed-by: Patrick Burke <patrick.burke@nokia.com>