Modify the sphere so it is a mesh.
There was a performance bug illustrated by a Qt3D user where an app that
had a large number of identical spheres, took minutes to load.

The problem was that even though the geometry was identical there was no
way to utilise this to create a scene with multiple items referencing
one sphere, because the geometry actually was being created in the draw
function....

Now fixed so that the Sphere becomes a SphereMesh which can be
referenced from any number of Item3D's.

Added a Sphere convenience class so that legacy code will keep working.

Updated the solarsystem_qml class to use the new functionality.

Change-Id: I62a60b519bbb048ecadaa09e0d5520ddb49ff100
Reviewed-by: Patrick Burke <patrick.burke@nokia.com>