Mesh conversion on Cylinder and Capsule.
Follow the example of the Sphere and convert Cylinder and Capsule to
being meshes as well.

Add some useful doc to the three classes showing the performance and
other considerations for each.

Add a basic manual test, with a README.

Slightly improve the warning message from QDeclarativeMesh.

Fix a bug with Sphere where nothing would display if the default
values for a sphere were used. This was because the createGeometry()
function was only triggered when properties changed.  Now also
trigger it if draw is called and the scenegraph for the object
has not yet been created.  Same code was also required on the
new shape meshes.

Change-Id: I475e34e31741ba8583587961fc3d0ee3d03bcb9d
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>