Make Qt3D buildable without widgets or opengl.
Pull out all references to QGLContext, and QGLWidget and so on; and
replace with the equivalent from the gui library.  Where there is no
equivalent, eg the BindOptions enum, then reimplement.

Also get rid of the harmattan and symbian packaging goop.  This had
become completely unmaintainable, and was making it very hard to see
what dependencies were being pull in by the various parts of the build
system.  In order to have any confidence that conditional compilation
would not add in some problematic dependency, clean all this up.

Task-number: QTBUG-23299
Change-Id: Ied92ca1397076986e455674b064d8540bb13c0df
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>