Work around openGLVersion bug.
A bug QTBUG-23504 in openGLVersion meant that we were getting a lot of
warnings about Cocoa resources not being released properly - memory
leaks.  Similar problems exist on other platforms possibly.

In any case the version checks can be deferred until a context is
available.

Change-Id: Id3a6872c5f91bae701f8fcfe35d656be6f74d2fd
Note: this is *not* a fix for QTBUG-23504, that bug is in Qt.
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>