Remove the graphics view example & support.
This is not supported in qt5 so remove it.

In the future we will encourage putting QML applications on the side of
a cube or teapot using the FBO support.  But doing this via QGraphics*
code which really belongs to both the 4.8 days and the old imperative
non-QML way of doing things - not a good idea.  We can't recommend that
as a viable approach.

Change-Id: I70f69f17edd949e313758f930bb75cdaad3c46c5
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>