Integrate Qt3D into Qt5 and fix some qdoc errors.
This doc change by 'integrating into Qt5' also made some changes to the
build system, so that now you need QT+=3d everywhere (instead of
QT+=qt3d as it was previously).

We have been asked to make this change, and it seems it is required for
this doc system to work, so lets get it done.

Note that the original change had a docsnippets target but that did not
work at all and will have be done in a later commit.

Change-Id: Icf4bc6b4be9494f1752e18ca1f5198c8fe8ecf53
Reviewed-by: Danny Pope <daniel.pope@nokia.com>