Suppress debug output on model load.
Switch to using an environment variable for displaying model debug
output.

The _DEBUG macro is not a guarantee that the debug output will only be
produced in debug builds, since its not respected by some compilers.
In any case, even in debug mode lots of spurious output is not desired.

Let's remove the _DEBUG macro, in favour of a documented environment
variable.

Change-Id: Ieb2f3647118e9380be75b2d3b22589a488bc4209
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>