Add examples doc & fix doc a bit.
Create new qdoc examples doc to reference our four showcase examples:
cube, lander, monkeygod and shaders.

There was an issue with the shaders tutorial because even tho' this is
under a seperate directory qdoc pulls them all into the same namespace
with the \example tag so this had to be changed to shadertut.

Also began fixing qdoc errors but there is a huge pile and this fix
needs to go in.

Task-number: QTBUG-24135
Change-Id: I940d88b264f2753811b5eabb6d1f787d47a172f0
Reviewed-by: Julian de Bhal <julian.debhal@nokia.com>
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>