Revert 00af107 to fix index pg clobber.
Looks like the assumption in 00af107 was wrong.  Seems as though the Qt
Module doc http://doc.qt.nokia.com/5.0-snapshot/modules.html wants to
point to qt3d at qml-qt3d.html so lets try that.  Anything for now so as
not to clobber the main index page for Qt5.

Task-number: QTBUG-24316
Change-Id: I9336649c1a45f8f00d6384b8f8d2a5982292f2eb
Reviewed-by: Alan Alpert <alan.alpert@nokia.com>