Revert 00af107 to fix index pg clobber.
Looks like the assumption in 00af107 was wrong.  Seems as though the Qt
Module doc http://doc.qt.nokia.com/5.0-snapshot/modules.html wants to
point to qt3d at qml-qt3d.html so lets try that.  Anything for now so as
not to clobber the main index page for Qt5.

Task-number: QTBUG-24316
Change-Id: I476a7655b207b64bef2b64f5b2aba893eea33373
Reviewed-by: Alan Alpert <alan.alpert@nokia.com>
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>