Implement suggested change to avoid clashes.
Note that this bug was quickly addressed by a previous change, but this
further change implements some code-review suggestions.

See the comments on http://codereview.qt-project.org/16528

Renames and changes the targets for this file to avoid clashes.  Note
that this change relies on another change in qt/qtdoc:

http://codereview.qt-project.org/16612

Task-number: QTBUG-24316
Change-Id: I22b5cbadaa2d09ce56f1799377e23b49f399b920
Reviewed-by: Sergey Dubitskiy <sergey.dubitskiy@nokia.com>
Reviewed-by: Alan Alpert <alan.alpert@nokia.com>
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>