Add a hacky script to help with qdoc errors.
Now that it is not practicable to build our documentation outside of the
qt5 doc build process we are in the situation of getting around 5000
lines (as of now) full of errors from qdoc over just the projects we
depend on.  This script tries to tame that output a bit and still keep
the useful messages.

Change-Id: I3cb0ea08c2cdd4527039534f74d37c13c400df90
Reviewed-by: Danny Pope <daniel.pope@nokia.com>
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>