Fix unit tests + builds on windows.
Some of these unit test problems were not just windows, but the issues
manifested most badly on windows.

Create a mock view object that is similar to the real QGLView, but more
specially designed for unit tests, and setup with sensible defaults.  Use
this for the main failing unit tests.

This fixed some of the issues, but there is also a problem with window
geometry on windows, and it is not going to be fixed as part of this change.
See https://bugreports.qt-project.org/browse/QTBUG-24539 for more.

Also fix a few other minor breaks including tutorials not being packaged
properly, and places where literal strings were used.

Change-Id: Id4f78823f851c822709208c3793f35a0dd472e5e
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>