Do not call glGetString() outside a OpenGL context
We can not call qt_has_uint_buffers() inside the ctor of
QGLIndexBufferPrivate, because at this point we are not inside a valid
OpenGL context, so the call to glGetString() will cause a SEGFAULT.

Postpone the call to qt_has_uint_buffers() until we are about to bind
the buffer in QGLIndexBuffer::bind() instead.

Also make the qt_uint_buffers() function thread safe.

Task-number: QTBUG-25175

Change-Id: Idd49798144b528dadd340e4355113a4bea14b6c9
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>
Reviewed-by: Sergey Dubitskiy <sergey.dubitskiy@nokia.com>