Fix qml3d tests somewhat.
There are still a number of problems here.  Have created new bugs for
those.

While on the job have also ported to the correct best-practice for
QtQuick tests which is to use the TESTDATA feature implemented in the
qmltestcase.prf feature file, in preference to the compile time macro.

Task-number: QTBUG-25044
Change-Id: Iff3046d6b910a03be001064f63275554f83459db
Reviewed-by: Julian de Bhal <julian.debhal@nokia.com>