Fix qml3d tests incl windows.
There are a lot of problems with our tests on platforms where the
fixed-function pipeline is in force, that is no shaders are available.
We have code paths that are supposed to deal with that but they're
poorly maintained and now crash.  Try to fix a few of them, and work
around others which are exposed by the qml3d tests on windows CI boxes
(that have software OpenGL 1.x).

We notice that QT_QPA_PLATFORM=minimal causes a crash with QML tests
that have a test function without a surrounding TestCase element.
Let's fix that and see if it helps also the mysterious windows test
crashing problem.

Task-number: QTBUG-25276
Change-Id: Ieeb67536f86467bb6819b62319ac3cee2219a8ae
Reviewed-by: Sarah Jane Smith <sarah.j.smith@nokia.com>