Handling of different OpenGL versions
3 Helper classes have been created:
- QGraphicsHelperES2 for OpenGL ES2 with the GL_OES_vertex_array_object
  extenstion for VAO
- QGraphicsHelperGL2 for OpenGL < 3.2 with the GL_ARB_vertex_array_object
  extension for VAO
- QGraphicsHelperGL3 for OpenGL >= 3.2

The helpers wraps optimized GL calls for each OpenGL version.
QGraphicsContext detects the highest supported OpenGL version
between 3.2, 2.0 and ES2 and instantiate the best matching helper.
All drawing commands that are requested to QGraphicsContext and then
directly sent to the helper.
Shaders version has been set to 150 to comply with OpenGL 3.2.

Different set of shaders with various versions should be bundled with Qt3D
later on.

Change-Id: Id618f1f4f0c2ca65204adedfffac26bdfc37d9f5
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>