FrameGraph and frontend FrameGraphNode
All framegraph building blocks are now subclasses of FrameGraphNode.
An Entity can reference the framegraph configuration to use by
by declaring inline a FrameGraphBuildingBlock or referencing it.

RenderSceneBuilder retrieves the first FrameGraph in the scene and builds
a tree of Render::FrameGraphNodes from the activeFrameGraph it references.
The Render::FrameGraphNodes are empty at the moment and will be filled in follow-up commmits

Change-Id: Id5f90421be16c3c979a6da2dccf19068bf7921ec
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>