Add enabled property to FrameGraphItem
The enabled property will be used by the code that traverses the active
framegraph each frame to decide if the sub-tree rooted at each node
should be traversed or not.

For example, this could be used to disable the subtree of a framegraph
responsible for rendering the geometry bounding volumes.

Change-Id: Ia264f443d8c7bf2d337a9a6624c3aa3c9bd0a48b
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>