RenderQueues of RenderView
RenderQueues class uses a QBounderCircularBuffer to store computed
RenderViews of a subset of the following few frames. This allows
for RenderViewJobs to compute future frames while the current frame
is being rendered. In case the RenderView submission takes too long,
the RenderView jobs will eventually be locked to allow the Renderer
to catch up and not miss any important state change.

This implementation tries to minimize the time needed to synchronize
the job threads with the rendering thread while providing thread safe
access to the queues if used correctly.

In RenderThread, the exec loop is replaced by a rendering loop and
the RenderQueue has been properly added to the Renderer.

Change-Id: I44f1f4065cdc7f39a8957a26114a905ea437834b
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>