RenderCommands fillings
Builds a set of RenderCommand for each RenderView.
The MeshData and worldTransformMatrix are set. The shader and
uniforms bindings aren't set at the moment. This should be done
in a follow up commit once I've properly figured out how to
associate a Shader to a Mesh a Material a RenderPass and an Effect
without relying on a RenderMaterial, RenderShader, RenderTechnique
so as to have a more data driven approach. A default shader and
uniform bindings should be sufficient for now.

Change-Id: I7a69ead2361b18a9d7b2554f090354f54d74303d
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>