Basic Rendering using the FrameGraph
Renderer is no longer a QObject. Therefore the RendererAspect can no longer
employ QMetaObject::invoke to call the setSurface method. As setSurface is now
called in the context of the RendererAspect/QAspectManager thread, the
QGraphicsContext instantiation can no longer take place in setSurface.
The QGraphicsContext instantiation is now performed in submit RenderView if it
appears that the Renderer's QGraphicsContext object is null. That ensures that
the QGraphicsContext, the QOpenGLContext and the QGraphicsHelper, which
initializes the GL functions, are now created in the RenderThread.

RenderView objects are now submitted in Renderer which initializes the scene
with the Viewport and Camera.

RenderCommands are submitted in the RenderThread and meshes are rendered using
a default Material and a default Technique.

Change-Id: Ic16833b860270bede25278d02475ff9b559fcaf2
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>