VAOManager
One VAO is created and associated to a MeshData and a RenderShader.
A ShaderManager was added but default Material / Technique and Shader
are still being used. Small changes to how the RenderCommands are executed
and the RenderView processed. Nice performance improvement overall.
Also proceeded with the removal of most of the hacked renderer.

Change-Id: Ie99bea7863e979e3dd88c8599472d527bb24b28c
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>