Settings RenderTechnique in RenderView
RenderShader are indexed by ShaderProgram Uuid.
Using the TechniqueFilters and RenderPassFilters, RenderView can now check if
the assigned Material has correct properties for Effect, Technique, RenderPass
and ShaderProgram.
For each of those but RenderPass, it builds a RenderEffect, RenderTechnique
and a RenderShader.
What is still to be determined is how and where uniforms have to be set. What
could be interesting would be to do that in RenderViewJobs and fill a
QUniformPack. That way, the Renderer could apply the uniforms later on during
submission without having to check the Material / Technique / Effect. Also it
would allow to sort RenderCommand instances according to the
QUniformPack. That would require setting the default material / technique
/ effect in the RenderViewJob rather than Renderer in case those set are not
valid.

Change-Id: Ie9e0cb7ef94e8bd64c503d26f3a327de23bede76
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>