FrameGraphItems C++ Qml API
FrameGraphItems are different from usual Qt3D elements as they inherit from
FrameGraphItem which is a sublcass of Node. As Qml items should now inherit
Quick3DNode instead of Node, this makes using old FrameGraphItems as they were
impossible because children cannot be added to a Node in Qml if the Node is
not a Quick3DNode.

One method that reduces code duplication to the minimum is creating abstract
classes for Viewport, RenderTargetSelector, CameraSelector, RenderPassFilter
and TechniqueFilter. Then the C++ FrameGraphItem Nodes can inherit from node
and one of those abstracts. On the QML side, Quick3D FrameGraphItems inherit
from Quick3DNode and one of the base class.

Because multiple inheritance of QObject is not possible, abstract
FrameGraphItems are not subclasses of QObject and are therefore declared as
Q_INTERFACES so as to be able to perform qobject_cast.

Change-Id: Ibf1350a53ce409ff03aad8d0fd8f5b7d6cc26fc0
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>