Quick3DEffect and EffectNode
Effect has been transformed to a non QObject abstract. EffectNode and
Quick3DEffect subclass Effect. This was needed so that Quick3DEffect can have
node children such as ShaderProgram.

Few changes to the backend and renderer to accommodate that a Material's effect
is now returned as a Node instead of an Effect as an Effect cannot be
registered as qml types as it has no QObject base.

Change-Id: Ief40f1d784151a085f244c35b3ed09a93dcc153e
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>