Quick3DScene and SceneNode
Examples updated to work with those new nodes. Quick3DScene and SceneNode
subclass Quick3DEntity, Scene and EntityNode, Scene respectively. This will
later ease the process of exposing Cameras loaded from a scene file to be used
by a CameraSelector in the FrameGraph. Also it allows to add children Node to
the scene on the Qml side.

Change-Id: I6843efd0aff60e51a9bd16c63ce19011afd5a8e5
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>