QQmlCustomParser for Shader properties
ShaderPropertyParser subclasses QQmlCustomParser. It allows to register Quick3DEffect
using qRegisterCustomType so that properties that are not parsed by the standard Qml
Engine can be parsed using the ShaderPropertyParser.

ShaderPropertyParser does nothing at the moment but show the type and name of
the properties. It will be completed later on when the declaration Uniforms
property from QML is determined.

Also Quick3DMaterial and Technique could be registered as CustomType using
ShaderPropertyParser as well so that we can define uniforms in Effect,
Material and Technique, Material overriding Technique and Technique overriding Effect.

Change-Id: I70d7b9a6112b929a71a814aaa44646881e78972f
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>