FrameGraphItems are observed for changes
The Backend FrameGraphNode observe when needed the frontend FrameGraphItem for
SceneChange notifications. This is the case when the camera property of a
CameraSelector is updated to reference another camera. This is also the case
for the Viewport rectangle which could be updated.

Will have to check later for TechniqueFilter/RenderPassFilter

Change-Id: Iaf592f6b2496790a11b05111def5d3af473a1e84
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>